Cleanup Redis module
Motivation:
There is some cleanup that can be done.

Modifications:
- Remove unused code
- Move common code to base class

Result:
Cleaner code.
