+++ codec-redis/src/main/java/io/netty/handler/codec/redis/ErrorRedisMessage.java
@@ -15,7 +15,6 @@
 
 package io.netty.handler.codec.redis;
 
-import io.netty.util.internal.StringUtil;
 import io.netty.util.internal.UnstableApi;
 
 /**
@@ -33,12 +32,4 @@ public final class ErrorRedisMessage extends AbstractStringRedisMessage {
         super(content);
     }
 
-    @Override
-    public String toString() {
-        return new StringBuilder(StringUtil.simpleClassName(this))
-                .append('[')
-                .append("content=")
-                .append(content())
-                .append(']').toString();
-    }
 }
